Fixed a refcount bug, do curl cleanup in ae thread
Using O instead of N in the Py_BuildObject was causing the
_acurl.Response object not to die. Once that was fixed it became
clean that the curl object would need to be locked if cleaned up
on the python thread so I've moved it back onto the ae curl thread.
More pipes were introduced to keep the code simple, each pipe now
has a single responsibility.